maximum projection is shown in the viewer window by switching to it with the press of 'E'. Switching back throws an error.

+++ src/main/java/bdv/cl/RenderSlice.java
@@ -293,7 +293,7 @@ public class RenderSlice {
 
 		long t = System.currentTimeMillis();
 		final RequiredBlocks requiredBlocks = getRequiredBlocks(sourceToScreen,
-				width, height, dimZ, new ViewId(timepointId, setupId));
+				width, height, (int) dimZ, new ViewId(timepointId, setupId));
 		t = System.currentTimeMillis() - t;
 		System.out.println("getRequiredBlocks: " + t + " ms");
 		t = System.currentTimeMillis();
